Avoid including non-billable work in report
The is_billable flag is on the task not the timeslip and it appears that
the reporting_type no longer (if it ever did) has any effect.

The nested=true query string means that the associated objects in each
timeslip are hydrated and so we can get immediate access to the task's
billable status withouth any extra API calls.